--- Assets/Prefabs/Boss.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 065046e213d17bd43b87ba2cda4c5e3e
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
