--- Assets/Prefabs/HardEnemy.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 33134d6e25696834e99948deef48ac9e
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
