+++ Assets/Prefabs/JerseyNumber.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 78038e5a69cd45749882ece9f3afb4d5
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
