--- Assets/Prefabs/MegaEnemy.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 9dc0ac3ad0a2ac440a70ff487b7fe66d
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
