--- Assets/Prefabs/SuperEnemy.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 08110d7e344982a41a7c02ca0d8134d6
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
