--- Assets/Prefabs/UltraEnemy.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: ec7e02bf4e11a0d47a1a95fc36535b6a
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
