--- Assets/Resources/Enemy.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 1bb1b8df667c38c4aaf60058e314b79b
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
