--- Assets/Scripts/EnemySpawner.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 39ad28a8560a74d4aabc704084a6f47e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
