--- Assets/Scripts/Player.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 781245e2875679b40a0f98048894f7e7
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
